refactor: reusable image upload
