+++ client/src/api/services/imgurService.js
@@ -1,0 +1,5 @@
+import api from '@/api/api';
+
+export const getImgurAccessToken = async () => {
+  return await api.get('/imgur/token');
+};
